♻️ move create_issue mutation in api crate
