+++ backend/api/src/application/github/mod.rs
@@ -1,0 +1,1 @@
+pub mod create_issue;
