+++ backend/api/src/application/mod.rs
@@ -1,4 +1,5 @@
 pub mod budget;
+pub mod github;
 pub mod payment;
 pub mod project;
 pub mod sponsor;
