+++ backend/github-proxy/src/domain/github/service.rs
@@ -49,13 +49,4 @@ pub trait Service: Send + Sync {
 		per_page: Option<u8>,
 		page: Option<u32>,
 	) -> Result<Vec<GithubIssue>>;
-
-	async fn create_issue(
-		&self,
-		repo_owner: &str,
-		repo_name: &str,
-		title: &str,
-		description: &str,
-		assignees: Vec<String>,
-	) -> Result<GithubIssue>;
 }
