+++ backend/github-proxy/src/infrastructure/github/service.rs
@@ -1,10 +1,6 @@
-use anyhow::anyhow;
-use derive_more::{From, Into};
-use domain::{GithubIssue, GithubIssueStatus, GithubIssueType, GithubRepositoryId};
-use infrastructure::github;
-use octocrab::models::issues::IssueStateReason;
+use domain::{GithubIssue, GithubRepositoryId};
+use infrastructure::github::{self, FromOctocrabIssue};
 use olog::{error, tracing::instrument};
-use thiserror::Error;
 
 use super::Contributors;
 use crate::domain::{
@@ -71,18 +67,20 @@ impl<P: github::OctocrabProxy> GithubService for P {
 		let octocrab_pull_requests = self.get_repository_PRs(repository_id).await?;
 		let pull_requests = octocrab_pull_requests
 			.into_iter()
-			.filter_map(|pr| match GithubProxyIssue::try_from(pr.clone()) {
-				Ok(pr) => Some(pr.into()),
-				Err(e) => {
-					error!(
-						error = e.to_string(),
-						repository_id = repository_id,
-						pullrequest_id = pr.id.0,
-						"Failed to process pull request"
-					);
-					None
+			.filter_map(
+				|pr| match GithubIssue::from_octocrab_pull_request(pr.clone()) {
+					Ok(pr) => Some(pr),
+					Err(e) => {
+						error!(
+							error = e.to_string(),
+							repository_id = repository_id,
+							pullrequest_id = pr.id.0,
+							"Failed to process pull request"
+						);
+						None
+					},
 				},
-			})
+			)
 			.collect();
 
 		Ok(pull_requests)
@@ -95,11 +93,9 @@ impl<P: github::OctocrabProxy> GithubService for P {
 		repo_name: &str,
 		issue_number: u64,
 	) -> GithubServiceResult<GithubIssue> {
-		let issue: GithubProxyIssue =
-			self.get_issue(repo_owner, repo_name, issue_number).await?.try_into().map_err(
-				|e: GithubIssueFromOctocrabResultError| GithubServiceError::Other(anyhow!(e)),
-			)?;
-		Ok(issue.into())
+		let issue = self.get_issue(repo_owner, repo_name, issue_number).await?;
+		GithubIssue::from_octocrab_issue(issue)
+			.map_err(|e| GithubServiceError::MissingField(e.to_string()))
 	}
 
 	#[instrument(skip(self))]
@@ -139,8 +135,8 @@ impl<P: github::OctocrabProxy> GithubService for P {
 			.search_issues(query, sort, order, per_page, page)
 			.await?
 			.into_iter()
-			.filter_map(|issue| match GithubProxyIssue::try_from(issue) {
-				Ok(issue) => Some(issue.into()),
+			.filter_map(|issue| match GithubIssue::from_octocrab_issue(issue) {
+				Ok(issue) => Some(issue),
 				Err(error) => {
 					error!(error = error.to_string(), "Failed to map Octocrab issue");
 					None
@@ -149,175 +145,10 @@ impl<P: github::OctocrabProxy> GithubService for P {
 			.collect();
 		Ok(issues)
 	}
-
-	#[instrument(skip(self))]
-	async fn create_issue(
-		&self,
-		repo_owner: &str,
-		repo_name: &str,
-		title: &str,
-		description: &str,
-		assignees: Vec<String>,
-	) -> GithubServiceResult<GithubIssue> {
-		let issue: GithubProxyIssue = self
-			.octocrab()
-			.issues(repo_owner, repo_name)
-			.create(title)
-			.body(description)
-			.assignees(assignees)
-			.send()
-			.await
-			.map_err(|e| GithubServiceError::Other(anyhow!(e)))?
-			.try_into()
-			.map_err(|e: GithubIssueFromOctocrabResultError| {
-				GithubServiceError::Other(anyhow!(e))
-			})?;
-		Ok(issue.into())
-	}
 }
 
 impl From<octocrab::models::User> for GithubUser {
 	fn from(user: octocrab::models::User) -> Self {
 		Self::new(user.id.0 as i32, user.login, user.avatar_url, user.html_url)
-	}
-}
-
-#[derive(From, Into)]
-struct GithubProxyIssue(GithubIssue);
-
-#[derive(Debug, Error)]
-pub enum GithubIssueFromOctocrabResultError {
-	#[error("Field '{0}' is not present")]
-	MissingField(String),
-	#[error(transparent)]
-	UnknownStatus(#[from] GithubIssueStatusFromOctocrabResultError),
-}
-
-impl TryFrom<octocrab::models::issues::Issue> for GithubProxyIssue {
-	type Error = GithubIssueFromOctocrabResultError;
-
-	fn try_from(issue: octocrab::models::issues::Issue) -> Result<Self, Self::Error> {
-		let id = issue
-			.id
-			.0
-			.try_into()
-			.expect("We cannot work with github ids superior to i32::MAX");
-
-		let number = issue
-			.number
-			.try_into()
-			.expect("We cannot work with github PR number superior to i32::MAX");
-
-		let issue_type = match issue.pull_request {
-			Some(_) => GithubIssueType::PullRequest,
-			None => GithubIssueType::Issue,
-		};
-
-		let status = get_status_from_issue(&issue)?;
-
-		Ok(domain::GithubIssue::new(
-			id,
-			number,
-			issue_type,
-			issue.title,
-			issue.html_url,
-			status,
-			issue.created_at,
-			issue.pull_request.and_then(|pr| pr.merged_at),
-			issue.closed_at,
-		)
-		.into())
-	}
-}
-
-impl TryFrom<octocrab::models::pulls::PullRequest> for GithubProxyIssue {
-	type Error = GithubIssueFromOctocrabResultError;
-
-	fn try_from(pull_request: octocrab::models::pulls::PullRequest) -> Result<Self, Self::Error> {
-		let id = pull_request
-			.id
-			.0
-			.try_into()
-			.expect("We cannot work with github ids superior to i32::MAX");
-
-		let number = pull_request
-			.number
-			.try_into()
-			.expect("We cannot work with github PR number superior to i32::MAX");
-
-		let title = pull_request
-			.title
-			.clone()
-			.ok_or_else(|| Self::Error::MissingField("title".to_string()))?;
-
-		let status = get_status_from_pull_request(&pull_request)?;
-
-		let created_at = pull_request
-			.created_at
-			.ok_or_else(|| Self::Error::MissingField("created_at".to_string()))?;
-
-		let html_url = pull_request
-			.html_url
-			.ok_or_else(|| Self::Error::MissingField("html_url".to_string()))?;
-
-		Ok(domain::GithubIssue::new(
-			id,
-			number,
-			GithubIssueType::PullRequest,
-			title,
-			html_url,
-			status,
-			created_at,
-			pull_request.merged_at,
-			pull_request.closed_at,
-		)
-		.into())
-	}
-}
-
-#[derive(Debug, Error)]
-pub enum GithubIssueStatusFromOctocrabResultError {
-	#[error("Unknown octocrab state: '{0}'")]
-	UnknownState(String),
-	#[error("Field '{0}' is not present")]
-	MissingField(String),
-}
-
-fn get_status_from_issue(
-	issue: &octocrab::models::issues::Issue,
-) -> Result<GithubIssueStatus, GithubIssueStatusFromOctocrabResultError> {
-	match issue.state {
-		octocrab::models::IssueState::Open => Ok(GithubIssueStatus::Open),
-		octocrab::models::IssueState::Closed =>
-			match issue.pull_request.as_ref().and_then(|pr| pr.merged_at) {
-				Some(_) => Ok(GithubIssueStatus::Merged),
-				None => match issue.state_reason {
-					Some(IssueStateReason::Completed) => Ok(GithubIssueStatus::Completed),
-					Some(IssueStateReason::NotPlanned) => Ok(GithubIssueStatus::Cancelled),
-					_ => Ok(GithubIssueStatus::Closed),
-				},
-			},
-		_ => Err(GithubIssueStatusFromOctocrabResultError::UnknownState(
-			format!("{:?}", issue.state),
-		)),
-	}
-}
-
-fn get_status_from_pull_request(
-	pull_request: &octocrab::models::pulls::PullRequest,
-) -> Result<GithubIssueStatus, GithubIssueStatusFromOctocrabResultError> {
-	let state = pull_request.state.as_ref().ok_or_else(|| {
-		GithubIssueStatusFromOctocrabResultError::MissingField("state".to_string())
-	})?;
-
-	match state {
-		octocrab::models::IssueState::Open => Ok(GithubIssueStatus::Open),
-		octocrab::models::IssueState::Closed => match pull_request.merged_at {
-			Some(_) => Ok(GithubIssueStatus::Merged),
-			None => Ok(GithubIssueStatus::Closed),
-		},
-		_ => Err(GithubIssueStatusFromOctocrabResultError::UnknownState(
-			format!("{:?}", state),
-		)),
 	}
 }
