+++ backend/github-proxy/src/presentation/graphql/error.rs
@@ -33,5 +33,3 @@ impl IntoFieldError for Error {
 		FieldError::new(msg, graphql_value!({ "reason": reason }))
 	}
 }
-
-pub type Result<T> = std::result::Result<T, Error>;
